{{ cookiecutter.project_slug }}/scripts/analytics_integration.py
#!/usr/bin/env python3
"""
Analytics integration for MkDocs knowledge base.
This script adds analytics tracking to MkDocs pages and provides
dashboard generation for content popularity tracking.
"""

import os
import re
import json
import logging
import argparse
from pathlib import Path

# Set up logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger("analytics_integration")

# Dashboard template for analytics visualization
DASHBOARD_TEMPLATE = """
// Knowledge Base Analytics Dashboard
import React, { useState, useEffect } from 'react';
import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, Legend, ResponsiveContainer } from 'recharts';

export const AnalyticsDashboard = () => {
  const [data, setData] = useState([]);
  const [loading, setLoading] = useState(true);
  const [error, setError] = useState(null);
  
  useEffect(() => {
    // In a real implementation, this would fetch data from your analytics API
    // This is just example data
    const mockData = [
      { page: 'Home', views: 1240, uniqueVisitors: 845 },
      { page: 'Getting Started', views: 890, uniqueVisitors: 654 },
      { page: 'User Guide', views: 760, uniqueVisitors: 542 },
      { page: 'API Reference', views: 680, uniqueVisitors: 478 },
      { page: 'Examples', views: 590, uniqueVisitors: 402 },
    ];
    
    setData(mockData);
    setLoading(false);
  }, []);
  
  if (loading) return <div>Loading analytics data...</div>;
  if (error) return <div>Error loading analytics: {error}</div>;
  
  return (
    <div>
      <h2>Knowledge Base Analytics</h2>
      <div style={{ width: '100%', height: 400 }}>
        <ResponsiveContainer>
          <BarChart data={data} margin={{ top: 20, right: 30, left: 20, bottom: 5 }}>
            <CartesianGrid strokeDasharray="3 3" />
            <XAxis dataKey="page" />
            <YAxis />
            <Tooltip />
            <Legend />
            <Bar dataKey="views" fill="#8884d8" name="Page Views" />
            <Bar dataKey="uniqueVisitors" fill="#82ca9d" name="Unique Visitors" />
          </BarChart>
        </ResponsiveContainer>
      </div>
    </div>
  );
};

export default AnalyticsDashboard;
"""

# Templates for different analytics systems
ANALYTICS_SYSTEMS = {
    "google-analytics": {
        "description": "Google Analytics (GA4)",
        "js_template": """
<!-- Matomo -->
<script>
  var _paq = window._paq = window._paq || [];
  _paq.push(['trackPageView']);
  _paq.push(['enableLinkTracking']);
  (function() {
    var u="//{{matomo_url}}/";
    _paq.push(['setTrackerUrl', u+'matomo.php']);
    _paq.push(['setSiteId', '{{site_id}}']);
    var d=document, g=d.createElement('script'), s=d.getElementsByTagName('script')[0];
    g.async=true; g.src=u+'matomo.js'; s.parentNode.insertBefore(g,s);
  })();
</script>
"""": """
<!-- Google Analytics -->
<script async src="https://www.googletagmanager.com/gtag/js?id={{measurement_id}}"></script>
<script>
  window.dataLayer = window.dataLayer || [];
  function gtag(){dataLayer.push(arguments);}
  gtag('js', new Date());

  gtag('config', '{{measurement_id}}');
</script>
"""
    },
    "plausible": {
        "description": "Plausible Analytics (privacy-friendly)",
        "js_template": """
<!-- Plausible Analytics -->
<script defer data-domain="{{domain}}" src="https://plausible.io/js/script.js"></script>
"""
    },
    "matomo": {
        "description": "Matomo (self-hosted)",
        "js_template