{{ cookiecutter.project_slug }}/requirements.txt
# MkDocs and themes
mkdocs>=1.4.0
mkdocs-material>=8.5.0
mkdocs-git-revision-date-localized-plugin>=1.1.0

# Enhanced features
mkdocs-minify-plugin>=0.6.4
mkdocs-simple-hooks>=0.1.5
mkdocs-glightbox>=0.3.5
mkdocs-material-extensions>=1.3.1
pymdown-extensions>=10.7

# Development & automation
watchdog>=2.1.9
PyYAML>=6.0
cookiecutter>=2.1.1
pytest>=7.0.0
black>=22.10.0

# Optional dependencies
htpasswd>=0.4.0  # For authentication
pillow>=9.4.0   # For social cards generation
